+++ app/lib/pages/progress_page.dart
@@ -68,7 +68,10 @@ class _ProgressPageState extends State<ProgressPage> with Refena {
 
       if (ref.read(settingsProvider).autoFinish) {
         _finishTimer = Timer.periodic(const Duration(seconds: 1), (timer) {
-          if (ref.read(progressProvider).getFinishedCount(widget.sessionId) == _selectedFiles.length) {
+          final finished = ref.read(serverProvider)?.session?.files.values.map((e) => e.status).isFinishedOrError ??
+              ref.read(sendProvider)[widget.sessionId]?.files.values.map((e) => e.status).isFinishedOrError ??
+              true;
+          if (finished) {
             if (_finishCounter == 1) {
               timer.cancel();
               exit();
@@ -194,6 +197,9 @@ class _ProgressPageState extends State<ProgressPage> with Refena {
       speedInBytes = null;
     }
 
+    final fileStatusMap = receiveSession?.files.map((k, f) => MapEntry(k, f.status)) ?? sendSession!.files.map((k, f) => MapEntry(k, f.status));
+    final finishedCount = fileStatusMap.values.where((s) => s == FileStatus.finished).length;
+
     return WillPopScope(
       onWillPop: () async {
         if (await _onWillPop() && mounted) {
@@ -274,7 +280,7 @@ class _ProgressPageState extends State<ProgressPage> with Refena {
                 final file = _files[index - 2];
                 final String fileName = receiveSession?.files[file.id]?.desiredName ?? file.fileName;
 
-                final fileStatus = receiveSession?.files[file.id]?.status ?? sendSession!.files[file.id]!.status;
+                final fileStatus = fileStatusMap[file.id]!;
                 final savedToGallery = receiveSession?.files[file.id]?.savedToGallery ?? false;
 
                 final String? filePath;
@@ -378,6 +384,17 @@ class _ProgressPageState extends State<ProgressPage> with Refena {
                             ],
                           ),
                         ),
+                        if (sendSession != null && fileStatus == FileStatus.failed)
+                          IconButton(
+                            icon: const Icon(Icons.refresh),
+                            onPressed: () async {
+                              await ref.notifier(sendProvider).sendFile(
+                                    sessionId: widget.sessionId,
+                                    file: sendSession.files[file.id]!,
+                                    isRetry: true,
+                                  );
+                            },
+                          ),
                       ],
                     ),
                   ),
@@ -418,7 +435,7 @@ class _ProgressPageState extends State<ProgressPage> with Refena {
                                 crossAxisAlignment: CrossAxisAlignment.start,
                                 children: [
                                   Text(t.progressPage.total.count(
-                                    curr: progressNotifier.getFinishedCount(widget.sessionId),
+                                    curr: finishedCount,
                                     n: _selectedFiles.length,
                                   )),
                                   Text(t.progressPage.total.size(