+++ common/lib/model/file_status.dart
@@ -7,3 +7,7 @@ enum FileStatus {
   failed,
   finished,
 }
+
+extension FileStatusIterable on Iterable<FileStatus> {
+  bool get isFinishedOrError => every((status) => const {FileStatus.skipped, FileStatus.failed, FileStatus.finished}.contains(status));
+}