+++ app/assets/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.15.4 (unreleased)
 
+- feat: add button to retry a failed file transfer (@Tienisto)
 - feat: show tooltip on the "Scan" button (@Tienisto)
 - feat: treat any URI as link, so it becomes clickable on receiver (e.g. file://, obsidian://) (@Tienisto)
 - feat(mobile): adjust padding between buttons in send tab to indicate that it's scrollable (@Tienisto)